+++ lib/app/data/datasources/api/endpoint_provider.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'genckoEndpoints.dart';
+import 'gencko_endpoints.dart';
 
 final getCoinsEndpointProvider = Provider((ref) {
   final dio = Dio(BaseOptions(
