--- lib/app/data/datasources/remote_datasource/endpoints/coinbase.dart
@@ -1,6 +1,0 @@
-// class Coinbase {
-//   static String baseUrl = "http://api.coinbase.com/v2/assets";
-//   static String getAllCoins = "$baseUrl/search?base=BRL";
-//   static String getCoinPrices(String coinId) =>
-//       "https://api.coingecko.com/api/v3/coins/bitcoin/market_chart?vs_currency=brl&days=1";
-// }
