--- lib/app/data/datasources/remote_datasource/endpoints/coingencko.dart
@@ -1,5 +1,0 @@
-// class Coingencko {
-//   static String baseUrl = "https://api.coingecko.com/api/v3";
-//   static String getAllCoins = "$baseUrl/list";
-//   static String getCoinById(String coinId) => "/coins/$coinId/tickersL";
-// }
