+++ lib/app/data/repositories/coin_repository_imp.dart
@@ -1,8 +1,7 @@
 import 'package:everest_crypto/app/data/datasources/api/models/coin_response.dart';
-import 'package:everest_crypto/app/domain/entities/coins_view_data.dart';
 import 'package:everest_crypto/app/domain/repositories/coin_repository.dart';
 
-import '../datasources/api/genckoEndpoints.dart';
+import '../datasources/api/gencko_endpoints.dart';
 import '../datasources/api/models/get_all_coins_response.dart';
 
 class CoinRepositoryImp implements ICoinRepository {
@@ -16,6 +15,4 @@ class CoinRepositoryImp implements ICoinRepository {
     return GetAllCoinsResponse(
         List.from(result.data.map((coin) => CoinResponse.fromJson(coin))));
   }
-
-  //Future<GetAllCoinByIdResponse> getCoins(Strin....
 }
