+++ lib/app/domain/repositories/coin_repository.dart
@@ -2,6 +2,4 @@ import '../../data/datasources/api/models/get_all_coins_response.dart';
 
 abstract class ICoinRepository {
   Future<GetAllCoinsResponse> getAllCoins(String vScurrency);
-
-  //Future<GetAllCoinByIdResponse> getCoins(Strin....
 }
