--- lib/app/domain/repositories/get_all_coins_repository.dart
@@ -1,5 +1,0 @@
-// import 'package:everest_crypto/app/domain/entities/coin_entity.dart';
-
-// abstract class GetAllCoinsRepository {
-//   Future<List<CoinEntity>> getAllCoins();
-// }
