+++ lib/app/domain/repositories/get_chart_config_repository.dart
@@ -2,6 +2,6 @@ import 'package:decimal/decimal.dart';
 
 import '../entities/chart_config_entity.dart';
 
-abstract class GetChartConfigRepository {
+abstract class IGetChartConfigRepository {
   ChartConfigEntity getChartConfig(List<Decimal> prices);
 }
