+++ lib/app/presenter/controllers/providers/chart_config_provider.dart
