+++ lib/app/presenter/controllers/providers/coin_prices_provider.dart
@@ -6,13 +6,6 @@ import '../../../data/repositories/coin_prices_repository_imp.dart';
 import '../../../domain/usecases/get_coin_prices_usecase.dart';
 import '../notifiers/coin_prices_notifier.dart';
 
-// final priceParameters = StateProvider<PricesResponse>(
-//   (ref) => PricesResponse(
-//       prices: Decimal.parse("0"),
-//       marketcaps: Decimal.parse("0"),
-//       totalvolumes: Decimal.parse("0")),
-// );
-
 final coinPricesRepositoryProvider = Provider((ref) {
   return CoinPricesRepositoryImp(ref.watch(getCoinsEndpointProvider));
 });
@@ -25,8 +18,3 @@ final coinsNotifierProvider =
     StateNotifierProvider<CoinPricesNotifier, AsyncValue<List<Decimal>>>((ref) {
   return CoinPricesNotifier(ref.watch(coinPricesUsecaseProvider));
 });
-
-// final coinPricesNotifierProvider =
-//     StateNotifierProvider<CoinPricesNotifier, AsyncValue<List<Decimal>>>((ref) {
-//   return CoinPricesNotifier(ref.watch(coinPricesUsecaseProvider));
-// });
