+++ lib/app/presenter/ui/portfolio/view/portfolio_page.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import '../../../../data/datasources/wallet_datasource.dart';
-import '../../../../domain/entities/coin_entity.dart';
 import '../../../controllers/providers/visible_provider.dart';
 import '../widgets/coin_list.dart';
 import '../widgets/wallet_details.dart';
